Changed float to double, short to long int, int to long int types throughout in GEMIPM2K and in the respective parts of GEMS-PSI (with DODs for MULTI data objects); replaced outdated calls (like memset(), memcpy() etc.) with loops, all this to improve compatibility with parallel architectures (SD)
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSTANDALONE/trunk@284 9ff7d5ee-7a19-0410-ab54-ec97d5369f81

+++ GEMIPM2K/m_param.h
@@ -1,5 +1,5 @@
 //-------------------------------------------------------------------
-// $Id: m_param.h 1069 2008-05-27 12:46:35Z gems $
+// $Id: m_param.h 1123 2008-11-25 13:23:26Z gems $
 //
 // Copyright (C) 2006,2007  S.Dmitrieva, D.Kulik
 //
@@ -36,7 +36,8 @@ extern const double R_CONSTANT, NA_CONSTANT, F_CONSTANT,
 
 struct BASE_PARAM
 { // Flags and thresholds for numeric modules
-    short PC,    // Mode of PhaseSelekt() (Selekt2) operation ( 0 1 2 ... )
+   long int 
+    PC,    // Mode of PhaseSelekt() (Selekt2) operation ( 0 1 2 ... )
     PD,  // Mode of DHH():0-invoke,1-at FIA only,2-last IPM it. 3-every IPM it.
     PRD, //Positive: mode GammaCalc(): 0-inactive, 1-FIA only, 2-SELEKT2 only 3-every IPM iteration
          // Negative: number of additional EFD-IPM improvement loops (-1 to -30 )
@@ -48,7 +49,7 @@ struct BASE_PARAM
     PE,  // Include electroneutrality condition? { 0 1 }
     IIM  // Maximum number of iterations in the main IPM algorithm { 500 }
     ;
-    double DG,   // Precision of LEM IPM solver (Jordan) { 1e-15 }
+  double DG,   // Precision of LEM IPM solver (Jordan) { 1e-15 }
     DHB, // Precision of mass balance deviations at EFD { 1e-8 }
     DS,  // Cutoff number of moles of a phase { 1e-10 }
     DK,  // Threshold of Dikin criterion of IPM convergence { 1e-4 }
@@ -112,7 +113,7 @@ public:
    void outMultiTxt( const char *path  );
    void readMulti( GemDataStream& ff );
    void readMulti( const char* path );
-   double calcMulti( int& PrecLoops_, int& NumIterFIA_, int& NumIterIPM_ );
+   double calcMulti( long int& PrecLoops_, long int& NumIterFIA_, long int& NumIterIPM_ );
 
 };
